+++ CHANGELOG.md
@@ -1,10 +1,18 @@
 
+# 2018/11/01 v1.4.0
+- Add 実行時間の見積もりを指定する属性( `m:15` で見積=15分の意)
+- Add "Report Today" (今日のタスク数と見積もり時間を表示)
+- Add バージョン情報表示機能( `python helper.py -v -i dummy` を実行)
+
+# 2018/10/11 v1.3.0
+- Add "Simple Completion" (現在行のみを補正(曜日修正など)する)
+
 # 2018/09/30 v1.2.1
 - Modify スキップの文法を `skip:月火水木金土日休平` 式に変更する
 - Modify 上記文法がハイライトされるようにする about trita.hilight
 
 # 2018/05/11 v1.2.0
-- Add タスク集計機能( `--report` オプション)
+- Add タスク集計機能( `python --report -i data.trita` を実行)
 
 # 2018/04/26 v1.1.0
 - Add "Walk +1 day"